% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coenocline}
\alias{coenocline}
\title{Simulate species abundance (counts) or occurrence along one or two
gradients}
\usage{
coenocline(x, responseModel = c("gaussian", "beta"), params,
  extraParams = NULL, countModel = c("poisson", "negbin", "bernoulli",
  "binary"), countParams = NULL, expectation = FALSE)
}
\arguments{
\item{x}{one of a numeric vector, a list with two components, each a numeric vector, or a matrix with two columns. The vectors are the locations along the gradient(s) at which species responses are to be simulated.}

\item{responseModel}{character; which species response model to use.}

\item{params}{a list of vectors each of which are parameters for the response model for each species. Alternatively, a matrix with one column per parameter and a row for each species.}

\item{extraParams}{a list containing additional parameters required for the response model. Examples include the correlation between gradients in the bivariate Gaussian response model. Components need to be named.}

\item{countModel}{character; if \code{expectation} is \code{FALSE}, the default, counts (occurrence) are generated using random deviates from the specified distribution.}

\item{countParams}{a list of additional parameters required to specify the distribution. An example is the parameter \eqn{\alpha}{alpha} in the negative binomial distribution. Components need to be named.}

\item{expectation}{logical; should the expectation (mean) response be returned (\code{TRUE})? If \code{FALSE} random counts or occurrences are generated using random draws from a suitably parameterised distribution, as stated in \code{countModel}.}
}
\value{
a matrix of simulated count or occurrence data, one row per gradient location, one column per species.
}
\description{
Simulate species abundance (counts) or occurrence along one or two gradients using well-known ecological response models and random draws from a Poisson, negative binomial or binomial distribution.
}
\details{
\code{coenocline()} is a generic interface to coenocline simulation allowing for easy extension and a consistent interface to a range of species response models and statistical distributions.

Two species response models are currently available; the Gaussian response and the generalized beta response model. Random count or occurrence data can be produced via random draws from a suitable distribution; in which case the values obtained from the specoes response function are used as the expectation of the distribution from which random draws are made.

Parameters for each species in the response model are supplied via argument \code{params} and can be provided in one of two ways: i) as a list with named components, each of which is a vector containing values for a single parameter for each species, or ii) as a matrix where each column contains the values for a single parameter and the rows represent species. In each case, the names of the list components or the column names of the matrix must be named for the arguments of the function implementing the species distribution model. See the examples.

Some species response models may require additional parameters not specified at the per species level. An example is the correlation between gradients in the bivariate Gaussian response model. Such parameters are passed via list \code{extraParams} and must be named accordingly so that they are passed to the corrct argument in the species response function.

The species response model defines the mean of expected response. (In the case of a species occurrence, the probability of occurrence is the expectation.) These represent paramterterised distributions. Random count or occurence data can be produced from these distributions by simulation from those distributions. In this case, a count or Bernoulli model is used and random draws from the distribution are made. Currently, the Poisson, the negative binomial, and the Bernoulli distribution are available. Some distributions may need additional parameters beyond the expectation; an example is the \eqn{\alpha}{alpha} parameter of (one parameterisation of) the negative binomial distribution. These parameters are specied via the list \code{countParams}.
}
\examples{
## Poisson counts along a single gradient, Gaussian response

x <- seq(from = 4, to = 6, length = 100)
opt <- c(3.75, 4, 4.55, 5, 5.5) + 0.5
tol <- rep(0.25, 5)
h <- rep(20, 5)

## set.seed(1)
y <- coenocline(x, responseModel = "gaussian",
                params = cbind(opt = opt, tol = tol, h = h),
                countModel = "poisson")
head(y)

y <- coenocline(x, responseModel = "gaussian",
                params = cbind(opt = opt, tol = tol, h = h),
                countModel = "poisson",
                expectation = TRUE)
matplot(x, y, type = "l", lty = "solid")


## Poisson counts along two correlated gradients, Gaussian response
set.seed(1)
x1 <- runif(300, min = 4, max = 6)
opt1 <- seq(4, 6, length = 5)
tol1 <- rep(0.25, 5)
x2 <- runif(300, min = 2, max = 20)
opt2 <- seq(2, 20, length = 5)
tol2 <- rep(1, 5)
h <- rep(20, 5)

set.seed(1)
params <- list(px = list(opt = opt1, tol = tol1, h = h),
               py = list(opt = opt2, tol = tol2))
y <- coenocline(cbind(x1, x2),
                responseModel = "gaussian",
                params = params,
                extraParams = list(corr = 0.5),
                countModel = "poisson")
head(y)

## Recreate Fig. 2 of Minchin (1987)
A0 <- c(5,4,7,5,9,8) * 10
m <- c(25,85,10,60,45,60)
r <- c(3,3,4,4,6,5) * 10
alpha <- c(0.1,1,2,4,1.5,1)
gamma <- c(0.1,1,2,4,0.5,4)
x <- 1:100
params <- list(m = m, A0 = A0, r = r, alpha = alpha, gamma = gamma)

## Expectations
set.seed(2)
y <- coenocline(x, responseModel = "beta",
                params = params,
                countModel = "poisson")
head(y)
matplot(y, type = "l", lty = "solid")

y <- coenocline(x, responseModel = "beta",
                params = params,
                countModel = "poisson", expectation = TRUE)
matplot(y, type = "l", lty = "solid")
}
\author{
Gavin L. Simpson
}
\keyword{datagen}

