% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.coenocline}
\alias{dim.coenocline}
\title{Return the Dimentions of the Simulated Count Matrix}
\usage{
\method{dim}{coenocline}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"coenocline"}, the result of a call to \code{\link{coenocline}}.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A vector of length 2 containing the number of rows and columns of the matrix of simulated counts.
}
\description{
Returns the number of rows and columns of the matrix of simulated counts.
}
\author{
Gavin L. Simpson
}
\seealso{
\code{\link{dim}}, \code{\link{coenocline}}
}

