% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distributions}
\alias{Binomial}
\alias{NegBin}
\alias{Poisson}
\alias{distributions}
\title{Wrappers to random number generators for use with coenocliner}
\usage{
NegBin(n, mu, alpha)

Poisson(n, mu)

Binomial(n, mu)
}
\arguments{
\item{n}{the number of random draws, equal to number of species times the number of gradient locations.}

\item{mu}{the mean or expectation of the distribution. For \code{Binomial} this is the probability of occurrence as given by the response function.}

\item{alpha}{numeric; parameter for the negative binomial distribution.}
}
\value{
a vector of random draws from the stated distribution.
}
\description{
Wrappers to random number generators for use with coenocliner
}
\author{
Gavin L. Simpson
}
\keyword{distribution}

