% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{head.coenocline}
\alias{head.coenocline}
\alias{tail.coenocline}
\title{Return the First of Last Part of an Object}
\usage{
\method{head}{coenocline}(x, ...)

\method{tail}{coenocline}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"coenocline"}, the result of a call to \code{\link{coenocline}}.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A matrix consisting of the first or last \code{n} rows of the matrix of simulated counts.
}
\description{
Returns the first of last \code{n} rows of the simulated counts.
}
\author{
Gavin L. Simpson
}
\seealso{
\code{\link[utils]{head}}, \code{\link{coenocline}}
}

