% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showParams}
\alias{showParams}
\title{List parameters of species response models}
\usage{
showParams(model = c("gaussian", "beta"))
}
\arguments{
\item{model}{character; the species response model for which parameters will be listed}
}
\value{
A character vector of parameters. The species response model is returned as attribute \code{"model"}. Attribute \code{"onlyx"} is a logical vector indicating which, if any, of the parameters are intended to be supplied only once per species and not for both gradients.
}
\description{
Returns the parameters of the indicated response model.
}
\examples{
showParams("gaussian")
}
\author{
Gavin L. Simpson
}
\seealso{
\code{\link{Gaussian}} and \code{\link{Beta}} for the species response model functions themselves.
}

