% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim1dBinom}
\alias{sim1dBinom}
\title{Simulate species occurrence along a single gradient}
\usage{
sim1dBinom(x, opt, tol, h, expectation = FALSE)
}
\arguments{
\item{x}{numeric; gradient locations.}

\item{opt}{numeric; species optima, one per taxon.}

\item{tol}{numeric; species tolerances, one per taxon.}

\item{h}{numeric between 0 and 1; species probability of presence at optima, one per taxon.}

\item{expectation}{logical; should expectations (probability of presence) rather than
discrete occurrence be returned?}
}
\value{
a matrix of simulated occurrences (if \code{expectation = FALSE}).
A matrix of probability of presence along the environmental gradient otherwise.
}
\description{
Simulates species occurrence along a single gradient
from a binomial distribution assuming a Gaussian response model.
}
\details{
The response model used is the classical Gaussian response with
parameters \eqn{\mu}{mu}, the species optimum, \eqn{t}, the
species tolerance, and \eqn{h}, the height of the response curve
at the optimum (see \code{\link{gaussianResponse}}).

If \code{expectation = FALSE} species occurrence (either presence or absence)
is drawn randomly from a binomial distribution with probability given
by the Gaussian response.
If \code{expectation = TRUE} then a continuous probability of presence (according to
the Gaussian response) is returned.
}
\section{Note}{

When called with \code{expectation = FALSE} the function does not use
the pseudorandom number generator, but when called with the defaults
a single call to pseudorandom number generator is made.
}
\examples{
## One species:
set.seed(1)
x <- seq(0, 30, length.out = 100)
Opt <- 18
Tol <- 3
H <- 0.7
y <- sim1dBinom(x, Opt, Tol, H, expectation = TRUE)
plot(x, y, main = "Probability of Presence", type="l")

y <- sim1dBinom(x, Opt, Tol, H, expectation = FALSE)
plot(x, y, main = "Occurrence")


## Multiple species
set.seed(1)
x1 <- seq(from = 4, to = 6, length = 100)
Opt <- c(3.75, 4, 4.55, 5, 5.5) + 0.5
Tol <- rep(0.25, 5)
H <- rep(0.7, 5)
y <- sim1dBinom(x1, Opt, Tol, H)

y <- sim1dBinom(x1, Opt, Tol, H, expectation = TRUE)
matplot(x1, y, type = "l", lty = "solid")
}
\author{
F. Rodriguez-Sanchez, based on code by Gavin L. Simpson
}
\seealso{
\code{\link{gaussianResponse}}
}

