% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulations}
\alias{simulations}
\title{Extract the Matrix of Simulated Counts}
\usage{
simulations(x, ...)
}
\arguments{
\item{x}{an object of class \code{"coenocline"}, the result of a call to \code{\link{coenocline}}.}

\item{...}{additional arguments passed to other methods. Currently ignored as the function is not and S3 generic.}
}
\value{
A matrix of simulated counts
}
\description{
Extracts the matrix of simulated counts from a suitable object.
}
\author{
Gavin L. Simpson
}

