% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{stack.coenocline}
\alias{stack.coenocline}
\title{Stack columns of a species coenocline simulation}
\usage{
\method{stack}{coenocline}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"coenocline"}}

\item{...}{arguments passed to other methods (not used).}
}
\description{
Stacks columns of a species coenocline simulation into long format suitable for use in statistical modeling or ggplot/lattice plots.
}
\seealso{
\code{\link{coenocline}}
}

