% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distributions}
\alias{Bernoulli}
\alias{BetaBinomial}
\alias{Binomial}
\alias{NegBin}
\alias{Poisson}
\alias{distributions}
\title{Wrappers to random number generators for use with coenocliner}
\usage{
NegBin(n, mu, alpha)

Poisson(n, mu)

Bernoulli(n, mu)

Binomial(n, mu, size)

BetaBinomial(n, mu, size, tau)
}
\arguments{
\item{n}{the number of random draws, equal to number of species times the number of gradient locations.}

\item{mu}{the mean or expectation of the distribution. For \code{Bernoulli}, \code{Binomial}, and \code{BetaBinomial()} this is the probability of occurrence as given by the response function.}

\item{alpha}{numeric; parameter for the negative binomial distribution.}

\item{size}{numeric; binomial denominator, the total number of individuals counted for example}

\item{tau}{numeric; the overdispersion parameter for the Beta-Binomial distribution. This is actually tau^2}
}
\value{
a vector of random draws from the stated distribution.
}
\description{
Wrappers to random number generators for use with coenocliner
}
\author{
Gavin L. Simpson
}
\keyword{distribution}

