% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expandBeta}
\alias{expandBeta}
\title{Expand gradient locations and species parameters of the
Generalise Beta response}
\usage{
expandBeta(x, m, A0, r, alpha, gamma)
}
\arguments{
\item{x}{numeric; gradient locations.}

\item{m}{numeric; species model abundance, one per taxon.}

\item{A0}{numeric; species abundance at \code{m}, one per taxon.}

\item{r}{numeric; range of \code{x} occupied by each species, one
per taxon.}

\item{alpha}{numeric; shape parameter of the generalised beta function.}

\item{gamma}{numeric; shape parameter of the generalised beta function.}
}
\value{
a matrix of Generalised Beta repsonse model parameters for
each gradient location. Specifically, a matrix with columns \code{x},
\code{m}, \code{A0}, \code{r}, \code{alpha}, and \code{gamma}.
}
\description{
An \code{\link{expand.grid}}-like utility function
for parameters of the Beta response model.
}
\details{
The response model used is the generalised Beta response with
the parameters: the position of the modal response (\code{m}), the
maximum abundance at the mode (\code{A0}), and the range of
occurence of the species on the gradient (\code{r}). In addition,
two positive shape parameters, \code{alpha} and \code{gamma}
constrain the shape of the response.\eqn{\mu}{mu}, the species
optimum, \eqn{t}, the species tolerance, and \eqn{h}, the height of
the response curve at the optimum.

\code{expandBeta} performs an \code{\link{expand.grid}}-like operation
on the gradient locations and species modela response, and then
replicates the remaining Beta response parameters to match the
replication of the modal locations.
}
\examples{
# Recreate Fig. 2 of Minchin (1987)
A0 <- c(5,4,7,5,9,8) * 10
m <- c(25,85,10,60,45,60)
r <- c(3,3,4,4,6,5) * 10
alpha <- c(0.1,1,2,4,1.5,1)
gamma <- c(0.1,1,2,4,0.5,4)
x <- 1:100

# expand parameter set
pars <- expandBeta(x, m, A0, r, alpha, gamma)
head(pars)

# Simulate expectation under Generalised Beta response
sim <- do.call(betaResponse, data.frame(pars))
sim <- matrix(sim, nrow = 100)

# plot
matplot(sim, type = "l", lty = "solid")
}
\author{
Gavin L. Simpson
}
\references{
Minchin P.R. (1987) Simulation of multidimensional
community patterns: towards a comprehensive model. \emph{Vegetatio}
\strong{71}, 145--156.
}

